Aquire new token if current token is about to expire
This fixes the edge case when the communication to vault is very slow
and the passed token is about to expire, it can happen that
the call to lookup self returns success but the following vault calls
still run into an error as the token just expired.

This change will fix this by calculating the maximal time, a call to
vault will take. This is capped to 60 seconds as the documentation
for the read and open timeout parameters were wrong. As the previous
documentation stated that the timout paramter are ms valued, they
are actually seconds. This could cause users to have high values
in these paramters. To prevent continues token refreshing in that case,
the maximal guaranteed token validity is capped to 60 seconds.